+++ src/components/Icons/index.tsx
@@ -9,3 +9,5 @@ export * from './Singletone/BellIcon';
 export * from './Singletone/QuestionIcon';
 export * from './Singletone/SearchIcon';
 export * from './Singletone/StarIcon';
+export * from './Singletone/ArrowIcon';
+export * from './Singletone/CalendarIcon';